updated account and contact
1. password checking, 8~30 length, containing letters, digits and symbols
2. login returning message is richer
3. readme: response code update
4. logout token detection
5. update contact/account does not require full information, see readme for more details

+++ account/views.py
@@ -11,13 +11,20 @@ from rest_framework.views import APIView
 from .models import ExtendedUser as User
 from common.code import *
 
+
 class LogOutViewSet(APIView):
 
     def post(self, request, *args, **kwargs):
         token = request.data.get('token')
-        Token.objects.filter(key=token).delete()
-        return JsonResponse({"code": ACCOUNT_LOGOUT_SUCCESS,
-                             "msg": "log out performed, token was disabled"})
+        try:
+            Token.objects.get(key=token).delete()
+            return JsonResponse({"code": ACCOUNT_LOGOUT_SUCCESS,
+                                 "msg": "log out performed, token was disabled"})
+        except:
+            return JsonResponse({"code": ACCOUNT_TOKEN_ERROR,
+                                 "msg": "token does not exist"})
+        # Token.objects.filter(key=token).delete()
+
 
 
 class UpdateViewSet(APIView):
@@ -28,16 +35,26 @@ class UpdateViewSet(APIView):
         first_name = request.data.get('first_name')
         last_name = request.data.get('last_name')
         dob = request.data.get('dob')
+        password = request.data.get('password')
+        email = request.data.get('email')
         try:
             user_id = Token.objects.get(key=token).user_id
         except:
             return JsonResponse({"code": ACCOUNT_TOKEN_ERROR,
-                                "msg": "token error"})
+                                 "msg": "token error"})
         user_obj = User.objects.get(id=user_id)
         # user_obj.email = email
-        user_obj.first_name = first_name
-        user_obj.last_name = last_name
-        user_obj.date_of_birth = dob
+        if first_name is not None:
+            user_obj.first_name = first_name
+        if last_name is not None:
+            user_obj.last_name = last_name
+        if dob is not None:
+            user_obj.date_of_birth = dob
+        if password is not None:
+            user_obj.password = password
+        if email is not None:
+            user_obj.email = email
+            user_obj.username = email
         try:
             user_obj.save()
         except Exception as e:
@@ -50,6 +67,15 @@ class UpdateViewSet(APIView):
 
 class RegisterViewSet(APIView):
 
+    def validate_password(self, password):
+        import re
+        reg = '^(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{8,30}$'
+        if re.match(reg, password):
+            return True
+        else:
+            raise Exception("Password must contains at least a digit, a letter, a upper case letter and a symbol, "
+                            "and length is between 8 and 30")
+
     def post(self, request, *args, **kwargs):
         # username = request.data.get('username')
         password = request.data.get('password')
@@ -57,7 +83,9 @@ class RegisterViewSet(APIView):
         first_name = request.data.get('first_name')
         last_name = request.data.get('last_name')
         try:
-            User.objects.create_user(username=email, password=password, email=email, first_name=first_name, last_name=last_name)
+            self.validate_password(password)
+            User.objects.create_user(username=email, password=password, email=email, first_name=first_name,
+                                     last_name=last_name)
         except Exception as e:
             return JsonResponse({"code": ACCOUNT_REGISTER_FAIL,
                                  "msg": str(e)})
@@ -67,13 +95,24 @@ class RegisterViewSet(APIView):
 
 class LoginViewSet(APIView):
 
+    def validate_user(self, request):
+        pass
+
     def post(self, request, *args, **kwargs):
         email = request.data.get('email')
         password = request.data.get('password')
+        if email is None:
+            return JsonResponse({"code": ACCOUNT_LOGIN_EMPTY_EMAIL,
+                                 "msg": "email is empty"})
+
+        if password is None:
+            return JsonResponse({"code": ACCOUNT_LOGIN_EMPTY_PASSWORD,
+                                 "msg": "password is empty"})
+
         user = auth.authenticate(username=email, password=password)
         if not user:
             return JsonResponse({"code": ACCOUNT_LOGIN_FAIL,
-                                 "msg": "wrong username/password"})
+                                 "msg": "wrong email/password"})
         Token.objects.filter(user=user).delete()
         return JsonResponse({"code": ACCOUNT_LOGIN_SUCCESS,
                              "msg": "login success",
